requirements.txt
chainlit 
python-dotenv 
-e . 
langchain-google-genai